+++ src/components/rootPage/Carousels/index.js
@@ -1,0 +1,4 @@
+export { default as Carousels } from '../Carousels';
+
+export { default as CarouselWithTitle } from './CarouselWithTitle';
+export { default as CarouselSkeleton } from './CarouselSkeleton';
