+++ src/components/rootPage/index.js
@@ -1,16 +1,12 @@
 export { Banner } from './Banner';
+export { CategoryPicker } from './CategoryPicker';
+export { Carousels } from './Carousels';
 
 export { default as Board } from './Board';
 export { Posters } from './Posters';
 export { default as Poster } from './Posters/Poster';
 
-export { CategoryPicker } from './CategoryPicker';
-
 export { default as Slide } from './Slide';
-export { default as Carousels } from './Carousels';
-export { default as CarouselWithTitle } from './CarouselWithTitle';
 
-export { default as CarouselSkeleton } from './CarouselSkeleton';
-
-export { default as MoreButton } from './CarouselSkeleton';
+export { default as MoreButton } from './Carousels/CarouselSkeleton';
 export { default as TopButton } from './TopButton';
